basic | File : Updated

+++ MarkPy/basic/filesystem.py
@@ -1,53 +1,37 @@
 from .logger import Logger
 
-from pathlib import Path
-
-# ParentPathCreationNewFileNotFound
-class FileException(Exception):
-    def __init__(self):
-        self.log.error(f"({self.name}): Parent path not exist of  {self.style.red(self.__file__)}")
-
-
-class File(Logger):
-    def __init__(self, name, path=None):
-        Atom.__init__(self, name)
-        self.__file__ = Path(path) / name if path else Path.cwd() / name
-        if not self.__file__.exists():
-            raise
-            with open(self.__file__, 'w') as fd:
-                fd.write('')
-        Logger.__init__(self, f'File.{self.__file__}', self.__file__.parent / f'{folder}.logs')
-        LogFile.__init__(self)
-        self.__log_file_created__()
-
-
-
-
-class LogFile():
-
-    def __log_file_created__(self):
-        self.log().debug(f"file {self.__file__} created by: {self.__name__}")
-
-
-
-class InitializationFile(Logger, LogFile):
-    def __init__(self, folder=None):
-        self.__file__ = Path(folder) if folder else Path.cwd()
-        if not self.__file__.exists():
-            with open(self.__file__, 'w') as fd:
-                fd.write('')
-        Logger.__init__(self, f'File.{self.__file__}', self.__file__.parent / f'{folder}.logs')
-        LogFile.__init__(self)
-        self.__log_file_created__()
-
-
-
-class File(InitializationFile):
-    pass
-
 
 from os import makedirs, remove
 from shutil import rmtree
+from pathlib import Path
+
+
+# ParentPathCreationNewFileNotFound
+class FileParentException(Exception):
+    def __init__(self, File):
+        File.log.error(f"({self.name}): Parent path not exist of  {File.red(File.__file__)}")
+
+
+class File(Logger):
+
+    __file_version__ = 1
+
+    def __init__(self, fileName, path=Path.cwd()):
+        Logger.__init__(self, fileName)
+        self.newLogAtom('File', self.__file_version__)
+        self.__file__ = Path(path) / fileName
+        init_opt = self.orange('attached')
+
+        if not Path(path).exists():
+            raise FileParentException(self)
+
+        if not self.__file__.exists():
+            init_opt = self.green('created')
+            with open(self.__file__, 'w') as fd:
+                fd.write('')
+
+        self.initialized()
+        self.log.debug(f' File {init_opt} -> {self.violet(self.__file__)}')
 
 
 class LogFolder():
